+++ spec/support/active_record_setup.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
-
 if testing_rails?
   # Remove the old sqlite database
   `rm -f ./_test_.db`
@@ -13,12 +12,24 @@ if testing_rails?
     ActiveRecord::Base.establish_connection(adapter: "jdbcsqlite3", database: "./_test_.db")
     SequelDB = Sequel.connect('jdbc:sqlite:./_test_.db')
   elsif ENV['DATABASE'] == 'POSTGRESQL'
-    ActiveRecord::Base.establish_connection(
+    ar_connection_options = {
+      host: "localhost",
       adapter: "postgresql",
       username: "postgres",
-      database: "graphql_ruby_test"
-    )
-    SequelDB = Sequel.connect('postgres://postgres:@localhost:5432/graphql_ruby_test')
+      password: ENV["PGPASSWORD"], # empty in development, populated for GH Actions
+      database: "graphql_ruby_test",
+    }
+    ActiveRecord::Base.establish_connection(ar_connection_options.merge(
+      database: "postgres"
+    ))
+    databases = ActiveRecord::Base.connection.execute("select datname from pg_database;")
+    test_db = databases.find { |d| d["datname"] == "graphql_ruby_test" }
+    if test_db.nil?
+      ActiveRecord::Base.connection.execute("create database graphql_ruby_test;")
+    end
+
+    ActiveRecord::Base.establish_connection(ar_connection_options)
+    SequelDB = Sequel.connect("postgres://postgres:#{ENV["PGPASSWORD"]}@localhost:5432/graphql_ruby_test")
   else
     ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: "./_test_.db")
     SequelDB = Sequel.sqlite("./_test_.db")