+++ javascript_client/CHANGELOG.md
@@ -1,5 +1,18 @@
 # graphql-ruby-client
 
+## 1.8.1
+
+- Sync: When `--url` is omitted, generate an outfile without syncing with a server
+
+## 1.8.0
+
+- Ably: Support server-side `cipher_base:` config in the client
+
+## 1.7.12
+
+- Ably: Add `rewind:` config so messages aren't lost between subscribe and registration of listener. #3210
+- Ably: Fix race condition where error was raised before the channel was available. #3210
+
 ## 1.7.11 (15 June 2020)
 
 - Ably: Improve channel state handling in case the initial subscription result contains errors #2993